requirements.txt
asgiref==3.2.3
black==21.5b2
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
flake8==3.9.2
gunicorn==20.0.4
idna==2.8
isort==5.8.0
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
whitenoise==5.0.1
